Fixed problem in the Function CountySalesTax in core.py
I wanted to broaden my test by returning 0.0 if the price was (0) or had a negative in it. The problem was that I did not put str around price. The error message that kept popping up was "Type float  is not iterable".

I plan to redo StateSalesTax function in core.py next.