Added Two Working Tests in test_core.py.
With Nate's help, I've made StateSalesTax and CountySalesTax in Core be available to pytest. I can feel it's only going to get harder from here, but I'll try my best.

I want to work on My future function Deposit in core, soon, and hopefully be able to write a test for it without help.