Final Push and Commit before leaving for weekend. Throwing errors.
There are errors and it's clearing the inventory. Copy and paste before trying it out again, so it does not completely empty it.
This is the error message:
File "/home/basecamp/rental_store_AlexandraFortner/core.py", line 89, in convert_back
    new_inventory += '\n{}, {:.2f}, {:.2f}'.format(item[0], item[1], item[2])
IndexError: string index out of range

I'll try to fix it when I come back.