Fixed last problem with revenue, made the math for the reciept live in core.py, and made a new function for my future option(return audio files).
I made id_letters_random, to make a random string of letters appear unique for every transaction.  My future plans for it are supposed to tie in with Returning audio files.  I have a general idea of how it will work, but applying it seems troublesome. Worried that I won't get done in time.

Working on the returning function is my next goal.