I am faced with a problem with return. Push before lunch.
In shell, the way to access returning is clear. I, however, don't know how to go about the function that connects to the shell  in disk.py. Line 69 on disk.py, I need to work on it.

After lunch, I need to finish working on disk.py's function( returning_update_history).