Fixed part of returning_audio_file() in shell.py.
If you put in the id number for your receipt, it will delete it from, the inventory, but I still have to make it to where it will return the specific deposit and return the rented inventory to the current inventory.

See above. Feel unsure on whether I will finish or not.